+++ pontos/copyright/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2020-2021 Greenbone Networks GmbH
+# Copyright (C) 2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -15,21 +15,6 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import os
+from .header import main
 
-from contextlib import contextmanager
-from pathlib import Path
-
-
-@contextmanager
-def use_cwd(path: Path) -> None:
-    """
-    Context Manager to change the current working directory temporaryly
-    """
-    current_cwd = Path.cwd()
-
-    os.chdir(str(path))
-
-    yield
-
-    os.chdir(str(current_cwd))
+__all__ = ['main']