+++ pontos/copyright/templates/GPL-3.0-or-later/template.sh
@@ -1,4 +1,5 @@
-# Copyright (C) 2020-2021 Greenbone Networks GmbH
+#!/bin/sh
+# Copyright (C) 2020 Company
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -14,22 +15,3 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-import os
-
-from contextlib import contextmanager
-from pathlib import Path
-
-
-@contextmanager
-def use_cwd(path: Path) -> None:
-    """
-    Context Manager to change the current working directory temporaryly
-    """
-    current_cwd = Path.cwd()
-
-    os.chdir(str(path))
-
-    yield
-
-    os.chdir(str(current_cwd))