+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
+- New command called pontos-copyright to update years in copyright headers [#58](https://github.com/greenbone/pontos/pull/58)
+
 ### Changed
 ### Deprecated
 ### Removed
@@ -14,13 +17,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/greenbone/pontos/compare/v0.3.1...HEAD
 
-
 ## [0.3.1] - 2021-01-05
+
 ### Added
 - add handling of PROJECT_DEV_VERSION in CMakeLists.txt if set [#32](https://github.com/greenbone/pontos/pull/32)
+
 ### Changed
 - set releasename to projectname version [#25](https://github.com/greenbone/pontos/pull/25)
 - separate signing tar and zipballs from release into a own command `sign` [#33](https://github.com/greenbone/pontos/pull/33)
+
 ### Fixed
 - project_dev handling was not working when there was a command after the set[#33](https://github.com/greenbone/pontos/pull/33)
 - use git-signing-key instead of signing-key on commit [42](https://github.com/greenbone/pontos/pull/42)